Componentize layout
