+++ src/pages/[slug].tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 import Image from "next/image";
-import { PageLayout } from "~/components/layout";
+import { PageLayout } from "~/components/layout/layout";
 import { LoadingPage } from "~/components/loading";
 import { api } from "~/utils/api";
