+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { useUser } from "@clerk/nextjs";
 
 import Feed from "~/components/feed";
-import { PageLayout } from "~/components/layout";
+import { PageLayout } from "~/components/layout/layout";
 import { CreatePostWizard } from "~/components/post-wizard";
 
 import { api } from "~/utils/api";
